+++ src/app/components/resources/CrearRecursos.jsx
@@ -4,7 +4,7 @@ import { ChevronLeft } from 'lucide-react';
 import { useRouter } from 'next/navigation';
 
 import TextInput from '@components/common/inputs/TextInput';
-import Dropdown from '@components/common/Dropdown';
+import Dropdown from '@/app/components/common/inputs/Dropdown';
 import FormContainer from '@components/common/FormContainer';
 import CustomButton from '@components/common/CustomButton';
 import Alert from '@components/common/alerts/Alert';
